Refactor: Consolidate imports in DashboardPage and MarketplacePage components
