+++ src/app/page.tsx
@@ -3,7 +3,6 @@ import { ROUTES, LOCALES } from '@/shared/config/constants';
 
 const { DASHBOARD } = ROUTES;
 
-
 export default function Home() {
   const locale = LOCALES.EN;
   redirect(`/${locale}/${DASHBOARD}`);
